requirements.txt
# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
moto==4.2.9

# Lambda dependencies
requests==2.31.0
python-dateutil==2.8.2

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
